requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1

# Agent Framework
langchain==0.1.12
langchain-community==0.0.28
langchain-groq==0.0.1
langsmith==0.1.22

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# Async & Queue
celery==5.3.6
redis==5.0.1

# Jira Integration
requests==2.31.0

# Vector Store
chromadb==0.4.24

# Dashboard
streamlit==1.31.1
plotly==5.18.0
pandas==2.2.0

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.26.0

# Utilities
pydantic==2.6.1
python-dateutil==2.8.2
structlog==24.1.0
